Moved test case to SparkDatasourceSuite
